# team_ontology/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OV enterprise ontology",
    "version": "sf.ontology.v1"
  },
  "paths": {
...
    "/api/v1/enterprise/ontology/feedback": {
      "get": {
        "tags": [
          "ontology-assets-v2"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_v1_enterprise_ontology_feedback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ontology"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_v1_enterprise_ontology_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/enterprise/ontology/access": {
      "put": {
        "tags": [
          "ontology-assets-v2"
        ],
        "summary": "Access",
        "operationId": "access_api_v1_enterprise_ontology_access_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessRequest": {
        "properties": {
          "subject": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          },
          "permissions": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "build",
                "approve",
                "publish",
                "observe",
                "feedback"
              ]
            },
            "type": "array",
            "maxItems": 6,
            "title": "Permissions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "subject",
          "permissions"
        ],
        "title": "AccessRequest"
      },
...
      "Feedback": {
        "properties": {
          "result_ref": {
            "type": "string",
            "title": "Result Ref"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result_ref",
          "note"
        ],
        "title": "Feedback"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}